feat: JspView 책임 분리
